backups-coms/coms/requirements/requirements.txt
boto3==1.10.28
botocore==1.13.28
certifi==2019.11.28
chardet==3.0.4
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2.7
-e git+https://github.com/django-admin-bootstrapped/django-admin-bootstrapped.git@30380f20cc82e898754e94b5d12cacafabca01bd#egg=django_admin_bootstrapped
django-anymail==7.0.0
django-appconf==1.0.3
django-apptemplates==1.4
django-authtools==1.7.0
django-bootstrap3==11.1.0
django-ckeditor==5.8.0
django-compat==1.0.15
django-compressor==2.3
django-extensions==2.2.5
django-filebrowser-no-grappelli==3.7.8
django-filter==2.2.0
django-floppyforms==1.8.0
django-form-utils==1.0.3
django-hijack==2.1.10
django-import-export==1.2.0
django-js-asset==1.2.2
django-permanent==1.1.6
django-rest-framework==0.1.0
django-select2==7.1.2
django-storages==1.8
django-tables2==2.2.1
django-webpack-loader==0.6.0
djangorestframework==3.10.3
docopt==0.4.0
docutils==0.15.2
et-xmlfile==1.0.1
freezegun==0.3.12
gunicorn==20.0.4
idna==2.8
jdcal==1.4.1
jmespath==0.9.4
lxml==4.4.2
mandrill==1.0.57
mandrill-for-py3==1.0.58
MarkupPy==1.14
MySQL-python==1.2.5
mysqlclient==1.4.1
odfpy==1.4.0
openpyxl==3.0.2
Pillow==6.2.1
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.1.2
rcssmin==1.0.6
requests==2.22.0
rjsmin==1.1.0
rules==2.1
s3transfer==0.2.1
six==1.13.0
spyne==2.13.4a1
sqlparse==0.3.0
suds-py3==1.3.4.0
tablib==0.14.0
tzlocal==1.2
unicodecsv==0.14.1
Unipath==1.1
urllib3==1.25.7
Werkzeug==0.16.0
whitenoise==4.1.4
xlrd==1.2.0
xlwt==1.3.0
xmltodict==0.12.0
